Fix non-legacy → legacy dependencies (Part I+II) and add dependency analysis doc
Part I+II from the dependency analysis:
- Move SendAgentMgmtCmd from legacy_agent_cmds.go to parentsync_cmds.go
- Move RunZoneList from legacy_agent_zone_cmds.go to zone_cmds.go
- Comment out pushKeystateInventoryToAllAgents in apihandler_funcs.go
- Comment out MP-specific branches in key_state_worker.go (mpremove
  transition, inventory push, weAreASigner check)

Add docs/2026-04-15-tdns-nonlegacy-to-legacy-dependency-analysis.md
documenting all non-legacy → legacy edges in tdns/v2.
