+++ v2/cli/zone_cmds.go
@@ -239,6 +239,32 @@ var zoneListCmd = &cobra.Command{
 	Run:   func(cmd *cobra.Command, args []string) { RunZoneList("auth", args) },
 }
 
+func RunZoneList(parent string, args []string) {
+	api, err := GetApiClient(parent, true)
+	if err != nil {
+		log.Fatalf("Error getting API client for %s: %v", parent, err)
+	}
+
+	cr, err := SendZoneCommand(api, tdns.ZonePost{
+		Command: "list-zones",
+	})
+	if err != nil {
+		fmt.Printf("Error from %q: %s\n", cr.AppName, err.Error())
+		os.Exit(1)
+	}
+
+	if cr.Msg != "" {
+		fmt.Printf("%s\n", cr.Msg)
+	}
+
+	switch tdns.Globals.Verbose {
+	case true:
+		VerboseListZone(cr)
+	case false:
+		ListZones(cr)
+	}
+}
+
 var zoneSerialBumpCmd = &cobra.Command{
 	Use:   "bump",
 	Short: "Bump SOA serial and epoch (if any) in tdns-auth version of zone",
